refactor and add func to liquidity
