+++ src/components/swap/ConfirmSwapModal.tsx
@@ -104,7 +104,7 @@ export default function ConfirmSwapModal({
       hash={txHash}
       content={confirmationContent}
       pendingText={pendingText}
-      trade={trade}
+      currency={trade?.outputAmount.currency}
     />
   )
 }
