+++ src/pages/RemoveLiquidity/index.tsx
@@ -488,6 +488,7 @@ export default function RemoveLiquidity({
             onDismiss={handleDismissConfirmation}
             attemptingTxn={attemptingTxn}
             hash={txHash ? txHash : ''}
+            currency={currencyB}
             content={() => (
               <ConfirmationModalContent
                 title={'You will receive'}
